Move graph.dot file to the top-level 'repo' directory
