+++ package.json
@@ -11,7 +11,9 @@
   },
   "scripts": {
     "install-dependencies": "pnpm install --strict-peer-dependencies=false",
-    "bootstrap": "pnpm install-dependencies && node ./scripts/link_packages_with_self.js",
+    "link-packages-with-self": "node ./scripts/link_packages_with_self.js",
+    "generate-packages-graph": "node ./scripts/generate_packages_graph.js ./repo",
+    "bootstrap": "pnpm install-dependencies && pnpm link-packages-with-self && pnpm generate-packages-graph",
     "build-env": "node packages/build-env/cli.js",
     "env:print": "pnpm build-env --print-env",
     "env:vars": "pnpm build-env --print-vars",
