+++ packages/build-env/package.json
@@ -15,7 +15,7 @@
     "url": "https://github.com/kiegroup/kie-tools.git"
   },
   "scripts": {
-    "print-build-env": "rimraf dist && node ./cli.js --print-cli-tools-check && node ./cli.js --print-env && node ./cli.js --print-config && node ./cli.js --build-graph",
+    "print-build-env": "rimraf dist && node ./cli.js --print-cli-tools-check && node ./cli.js --print-env && node ./cli.js --print-config",
     "build:dev": "pnpm print-build-env",
     "build:prod": "pnpm print-build-env"
   },
